btbs-fastapi-master/requirements.txt
aiohttp==3.9.3
aiokafka==0.10.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
astroid==3.1.0
async-timeout==4.0.3
asyncio-redis==0.16.0
asyncpg==0.29.0
attrs==23.2.0
click==8.1.7
dill==0.3.8
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
isort==5.13.2
mccabe==0.7.0
multidict==6.0.5
mysql-connector-python==8.3.0
packaging==23.2
permit-broadcaster==0.2.5
platformdirs==4.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.36.3
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==10.4
yarl==1.9.4
